requirements.txt
# API and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
aiohttp==3.8.6
httpx==0.25.1

# LLM Integration
openai==1.2.3
langchain==0.0.335
tiktoken==0.5.1

# Web Crawling
selenium==4.15.2
beautifulsoup4==4.12.2
playwright==1.39.0
scrapy==2.11.0

# Data Processing
pandas==2.1.2
numpy==1.26.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3

# Frontend
flask==2.3.3
jinja2==3.1.2

# Utilities
tqdm==4.66.1
colorama==0.4.6
